Split window helpers into focused files
