+++ internal/mux/spiral.go
@@ -260,82 +260,3 @@ func rootLeafCount(cell *LayoutCell) int {
 func spiralLayoutError(count int) error {
 	return fmt.Errorf("add-pane requires a canonical spiral layout prefix for %d panes", count)
 }
-
-func (w *Window) splitSubtreeRootWithOptions(root *LayoutCell, dir SplitDir, newPane *Pane, insertFirst bool, opts SplitOptions) (*Pane, error) {
-	if root == nil {
-		return nil, fmt.Errorf("no layout")
-	}
-	if w.ZoomedPaneID != 0 && !opts.KeepFocus {
-		w.Unzoom()
-	}
-
-	newLeaf := NewLeaf(newPane, 0, 0, 0, 0)
-
-	if !root.IsLeaf() && root.Dir == dir {
-		newLeaf.Parent = root
-		if insertFirst {
-			root.Children = append([]*LayoutCell{newLeaf}, root.Children...)
-		} else {
-			root.Children = append(root.Children, newLeaf)
-		}
-		root.distributeEqual()
-	} else {
-		oldRoot := root
-		parent := oldRoot.Parent
-		parentIdx := oldRoot.IndexInParent()
-
-		newRoot := &LayoutCell{
-			X: oldRoot.X, Y: oldRoot.Y, W: oldRoot.W, H: oldRoot.H,
-			Dir: dir,
-		}
-		if insertFirst {
-			newRoot.Children = []*LayoutCell{newLeaf, oldRoot}
-		} else {
-			newRoot.Children = []*LayoutCell{oldRoot, newLeaf}
-		}
-		newLeaf.Parent = newRoot
-		oldRoot.Parent = newRoot
-
-		if dir == SplitVertical {
-			secondW := (oldRoot.W - 1) / 2
-			firstW := oldRoot.W - 1 - secondW
-			if insertFirst {
-				newLeaf.W = firstW
-				newLeaf.H = oldRoot.H
-				oldRoot.ResizeSubtree(secondW, oldRoot.H)
-			} else {
-				newLeaf.W = secondW
-				newLeaf.H = oldRoot.H
-				oldRoot.ResizeSubtree(firstW, oldRoot.H)
-			}
-		} else {
-			secondH := (oldRoot.H - 1) / 2
-			firstH := oldRoot.H - 1 - secondH
-			if insertFirst {
-				newLeaf.W = oldRoot.W
-				newLeaf.H = firstH
-				oldRoot.ResizeSubtree(oldRoot.W, secondH)
-			} else {
-				newLeaf.W = oldRoot.W
-				newLeaf.H = secondH
-				oldRoot.ResizeSubtree(oldRoot.W, firstH)
-			}
-		}
-
-		newRoot.Parent = parent
-		if parent == nil {
-			w.Root = newRoot
-		} else {
-			parent.Children[parentIdx] = newRoot
-		}
-		root = newRoot
-	}
-
-	w.Root.FixOffsets()
-	w.resizePTYs()
-	w.restoreZoomedPaneSize()
-	if !opts.KeepFocus {
-		w.setActive(newPane)
-	}
-	return newPane, nil
-}
