+++ internal/mux/window.go
@@ -2,7 +2,6 @@ package mux
 
 import (
 	"fmt"
-	"sync/atomic"
 
 	"github.com/weill-labs/amux/internal/debugowner"
 )
@@ -279,451 +278,6 @@ func (w *Window) ClosePane(paneID uint32) error {
 	return nil
 }
 
-// Resize adjusts the layout to fit new terminal dimensions.
-func (w *Window) Resize(width, height int) {
-	w.assertOwner("Resize")
-	w.Width = width
-	w.Height = height
-	w.Root.ResizeAll(width, height)
-
-	w.resizePTYs()
-	w.restoreZoomedPaneSize()
-}
-
-// activePointCounter is a package-level monotonic counter for pane focus recency.
-var activePointCounter uint64
-
-// setActive updates the active pane and increments ActivePoint for recency tracking.
-func (w *Window) setActive(p *Pane) {
-	w.ActivePane = p
-	p.ActivePoint = atomic.AddUint64(&activePointCounter, 1)
-}
-
-// FocusPane sets the active pane directly (by pointer) and updates recency.
-// Used by the server when focusing by name or ID.
-// Auto-unzooms if a pane is zoomed and the target is a different pane.
-func (w *Window) FocusPane(p *Pane) {
-	w.assertOwner("FocusPane")
-	if w.ZoomedPaneID != 0 && p.ID != w.ZoomedPaneID {
-		w.Unzoom()
-	}
-	w.setActive(p)
-}
-
-// Focus changes the active pane. Direction is "next", "left", "right", "up", "down".
-// Uses tmux-style adjacency + perpendicular overlap + wrapping + recency tiebreaker.
-// Auto-unzooms if a pane is zoomed.
-func (w *Window) Focus(direction string) {
-	w.assertOwner("Focus")
-	if w.ZoomedPaneID != 0 {
-		w.Unzoom()
-	}
-	panes := w.Panes()
-	if len(panes) <= 1 {
-		return
-	}
-
-	if direction == "next" {
-		for i, p := range panes {
-			if p.ID == w.ActivePane.ID {
-				w.setActive(panes[(i+1)%len(panes)])
-				return
-			}
-		}
-		return
-	}
-
-	activeCell := w.Root.FindPane(w.ActivePane.ID)
-	if activeCell == nil {
-		return
-	}
-
-	// Try adjacent panes, then wrap to opposite edge.
-	best := w.findDirectional(activeCell, direction, false)
-	if best == nil {
-		best = w.findDirectional(activeCell, direction, true)
-	}
-
-	if best != nil {
-		w.setActive(best.Pane)
-	}
-}
-
-// findDirectional finds the best pane in the given direction from activeCell.
-// If wrap is true, searches from the opposite window edge instead.
-//
-// The algorithm checks two things for each candidate pane:
-//   - Adjacency: the candidate's edge touches the active pane's edge (with 1-cell border)
-//   - Perpendicular overlap: the candidate shares some range along the other axis
-//
-// Among matching candidates, the most recently active pane wins (recency tiebreaker).
-func (w *Window) findDirectional(activeCell *LayoutCell, direction string, wrap bool) *LayoutCell {
-	// vertical means we're moving along the Y axis (up/down).
-	// checkNear means adjacency is checked against the candidate's near edge
-	// (Y for down, X for right) rather than its far edge (Y+H for up, X+W for left).
-	vertical := direction == "up" || direction == "down"
-	checkNear := direction == "down" || direction == "right"
-
-	// Compute the edge that candidates must be adjacent to, and the
-	// perpendicular range they must overlap with.
-	var edge, rangeStart, rangeEnd int
-	switch direction {
-	case "up":
-		edge = activeCell.Y
-		if wrap {
-			edge = w.Height + 1
-		}
-	case "down":
-		edge = activeCell.Y + activeCell.H + 1
-		if wrap {
-			edge = 0
-		}
-	case "left":
-		edge = activeCell.X
-		if wrap {
-			edge = w.Width + 1
-		}
-	case "right":
-		edge = activeCell.X + activeCell.W + 1
-		if wrap {
-			edge = 0
-		}
-	}
-	if vertical {
-		rangeStart = activeCell.X
-		rangeEnd = activeCell.X + activeCell.W
-	} else {
-		rangeStart = activeCell.Y
-		rangeEnd = activeCell.Y + activeCell.H
-	}
-
-	var best *LayoutCell
-	var bestActivePoint uint64
-
-	w.Root.Walk(func(cell *LayoutCell) {
-		if cell.Pane == nil || cell.Pane.ID == w.ActivePane.ID {
-			return
-		}
-
-		// Check adjacency: candidate's edge must be exactly at our edge.
-		var candEdge, candStart, candEnd int
-		if vertical {
-			candStart = cell.X
-			candEnd = cell.X + cell.W
-			if checkNear {
-				candEdge = cell.Y
-			} else {
-				candEdge = cell.Y + cell.H + 1
-			}
-		} else {
-			candStart = cell.Y
-			candEnd = cell.Y + cell.H
-			if checkNear {
-				candEdge = cell.X
-			} else {
-				candEdge = cell.X + cell.W + 1
-			}
-		}
-		if candEdge != edge {
-			return
-		}
-
-		// Check perpendicular overlap (half-open interval intersection).
-		if candStart >= rangeEnd || candEnd <= rangeStart {
-			return
-		}
-
-		// Tiebreaker: most recently active pane wins.
-		if best == nil || cell.Pane.ActivePoint > bestActivePoint {
-			best = cell
-			bestActivePoint = cell.Pane.ActivePoint
-		}
-	})
-
-	return best
-}
-
-// PaneContentHeight returns the PTY height for a pane in a layout cell,
-// accounting for the per-pane status line.
-func PaneContentHeight(cellH int) int {
-	h := cellH - StatusLineRows
-	if h < 1 {
-		h = 1
-	}
-	return h
-}
-
-// ResizeBorder moves a border at position (x, y) by delta cells.
-// For vertical borders (vertical split), delta is applied horizontally.
-// For horizontal borders (horizontal split), delta is applied vertically.
-// Returns true if a resize was performed.
-func (w *Window) ResizeBorder(x, y, delta int) bool {
-	w.assertOwner("ResizeBorder")
-	hit := w.Root.FindBorderNear(x, y)
-	if hit == nil || delta == 0 {
-		return false
-	}
-
-	var leftSize, rightSize *int
-	if hit.Dir == SplitVertical {
-		leftSize = &hit.Left.W
-		rightSize = &hit.Right.W
-	} else {
-		leftSize = &hit.Left.H
-		rightSize = &hit.Right.H
-	}
-
-	// Clamp delta so neither side goes below PaneMinSize
-	if delta > 0 && *rightSize-delta < PaneMinSize {
-		delta = *rightSize - PaneMinSize
-	}
-	if delta < 0 && *leftSize+delta < PaneMinSize {
-		delta = -(*leftSize - PaneMinSize)
-	}
-	if delta == 0 {
-		return false
-	}
-
-	*leftSize += delta
-	*rightSize -= delta
-
-	// Propagate size changes to subtrees
-	if !hit.Left.IsLeaf() {
-		hit.Left.ResizeSubtree(hit.Left.W, hit.Left.H)
-	}
-	if !hit.Right.IsLeaf() {
-		hit.Right.ResizeSubtree(hit.Right.W, hit.Right.H)
-	}
-
-	w.Root.FixOffsets()
-	w.resizePTYs()
-	return true
-}
-
-// ResizeActive moves the nearest border in the given direction by delta cells,
-// following tmux's resize-pane semantics. The direction specifies which way the
-// border moves, not which way the pane grows.
-// direction is "left", "right", "up", or "down".
-// Returns true if a resize was performed.
-func (w *Window) ResizeActive(direction string, delta int) bool {
-	w.assertOwner("ResizeActive")
-	if w.ActivePane == nil {
-		return false
-	}
-	return w.ResizePane(w.ActivePane.ID, direction, delta)
-}
-
-// ResizePane resizes a specific pane by moving its nearest border in the given direction.
-// direction is "left", "right", "up", or "down". delta is the number of cells to move.
-// Returns true if a resize was performed.
-func (w *Window) ResizePane(paneID uint32, direction string, delta int) bool {
-	w.assertOwner("ResizePane")
-	if delta <= 0 {
-		return false
-	}
-	if w.IsLeadPane(paneID) {
-		return false
-	}
-	if w.ZoomedPaneID != 0 {
-		w.Unzoom()
-	}
-
-	// Map direction to split axis and change sign.
-	// Positive change grows the left/top sibling (border moves right/down).
-	// Negative change shrinks it (border moves left/up).
-	var axis SplitDir
-	var change int
-	switch direction {
-	case "left":
-		axis, change = SplitVertical, -delta
-	case "right":
-		axis, change = SplitVertical, delta
-	case "up":
-		axis, change = SplitHorizontal, -delta
-	case "down":
-		axis, change = SplitHorizontal, delta
-	default:
-		return false
-	}
-
-	leaf := w.Root.FindPane(paneID)
-	if leaf == nil {
-		return false
-	}
-
-	// Walk up the tree to find the nearest ancestor with matching axis.
-	cell := leaf
-	for cell.Parent != nil {
-		if cell.Parent.Dir == axis {
-			idx := cell.IndexInParent()
-			siblings := cell.Parent.Children
-			if len(siblings) < 2 {
-				return false
-			}
-
-			// tmux convention: resize the border adjacent to this cell.
-			// If we're the last child, use the border to our left (idx-1, idx).
-			// Otherwise, use the border to our right (idx, idx+1).
-			if idx == len(siblings)-1 {
-				idx--
-			}
-			if idx < 0 || idx+1 >= len(siblings) {
-				return false
-			}
-
-			var moved int
-			if change > 0 {
-				moved = w.resizePaneGrow(siblings, idx, axis, change)
-			} else {
-				moved = w.resizePaneShrink(siblings, idx, axis, -change)
-			}
-			if moved == 0 {
-				return false
-			}
-
-			w.Root.FixOffsets()
-			w.resizePTYs()
-			return true
-		}
-		cell = cell.Parent
-	}
-
-	return false
-}
-
-// Equalize rebalances the current logical root. When widths is true, root-level
-// columns are redistributed evenly. When heights is true, each logical column's
-// top-level rows are redistributed evenly. Returns true if the layout changed.
-func (w *Window) Equalize(widths, heights bool) bool {
-	w.assertOwner("Equalize")
-	if w.Root == nil || (!widths && !heights) {
-		return false
-	}
-
-	logical := w.logicalRoot()
-	if logical == nil {
-		return false
-	}
-
-	widthChanged := widths &&
-		!logical.IsLeaf() &&
-		logical.Dir == SplitVertical &&
-		len(logical.Children) > 1 &&
-		logical.equalizeChildrenNeeded()
-
-	columns := []*LayoutCell{logical}
-	if !logical.IsLeaf() && logical.Dir == SplitVertical {
-		columns = logical.Children
-	}
-
-	heightChanged := false
-	if heights {
-		for _, column := range columns {
-			if column == nil || column.IsLeaf() || column.Dir != SplitHorizontal || len(column.Children) < 2 {
-				continue
-			}
-			if column.equalizeChildrenNeeded() {
-				heightChanged = true
-				break
-			}
-		}
-	}
-
-	if !widthChanged && !heightChanged {
-		return false
-	}
-
-	if w.ZoomedPaneID != 0 {
-		w.Unzoom()
-	}
-
-	if widthChanged {
-		logical.distributeEqual()
-	}
-	if heights {
-		for _, column := range columns {
-			if column == nil || column.IsLeaf() || column.Dir != SplitHorizontal || len(column.Children) < 2 {
-				continue
-			}
-			if !column.equalizeChildrenNeeded() {
-				continue
-			}
-			column.distributeEqual()
-		}
-	}
-
-	w.Root.FixOffsets()
-	w.resizePTYs()
-	return true
-}
-
-func (w *Window) resizePaneGrow(siblings []*LayoutCell, idx int, axis SplitDir, needed int) int {
-	grower := siblings[idx]
-
-	// Match tmux layout_resize_pane_grow: walk tail-ward first, then fall back
-	// to the head if no right/bottom sibling can donate enough space.
-	remaining := w.transferSiblingRange(grower, siblings, idx, axis, needed, idx+1, len(siblings), 1)
-	if remaining == 0 {
-		return needed
-	}
-	remaining = w.transferSiblingRange(grower, siblings, idx, axis, remaining, idx-1, -1, -1)
-	return needed - remaining
-}
-
-func (w *Window) resizePaneShrink(siblings []*LayoutCell, idx int, axis SplitDir, needed int) int {
-	// Match tmux layout_resize_pane_shrink: grow the sibling across the border
-	// and walk left/up from the border cell looking for donors.
-	return needed - w.transferSiblingRange(siblings[idx+1], siblings, idx+1, axis, needed, idx, -1, -1)
-}
-
-func transferEdges(growerIdx, donorIdx int) (resizeEdge, resizeEdge) {
-	if donorIdx < growerIdx {
-		return resizeFromStart, resizeFromEnd
-	}
-	return resizeFromEnd, resizeFromStart
-}
-
-func (w *Window) transferSiblingRange(grower *LayoutCell, siblings []*LayoutCell, growerIdx int, axis SplitDir, remaining, start, stop, step int) int {
-	for donorIdx := start; donorIdx != stop && remaining > 0; donorIdx += step {
-		growerEdge, donorEdge := transferEdges(growerIdx, donorIdx)
-		remaining -= transferAxisSize(grower, siblings[donorIdx], axis, remaining, growerEdge, donorEdge)
-	}
-	return remaining
-}
-
-func transferAxisSize(grower, donor *LayoutCell, axis SplitDir, needed int, growerEdge, donorEdge resizeEdge) int {
-	available := donor.resizeCheck(axis)
-	if available == 0 {
-		return 0
-	}
-	if available > needed {
-		available = needed
-	}
-
-	grower.resizeToAxisFromEdge(axis, grower.axisSize(axis)+available, growerEdge)
-	donor.resizeToAxisFromEdge(axis, donor.axisSize(axis)-available, donorEdge)
-	return available
-}
-
-// resizePTYs resizes all pane PTYs to match their layout cell dimensions.
-func (w *Window) resizePTYs() {
-	w.Root.Walk(func(c *LayoutCell) {
-		if c.Pane != nil {
-			c.Pane.Resize(c.W, PaneContentHeight(c.H))
-		}
-	})
-}
-
-func (w *Window) restoreZoomedPaneSize() {
-	if w.ZoomedPaneID == 0 {
-		return
-	}
-	cell := w.Root.FindPane(w.ZoomedPaneID)
-	if cell != nil && cell.Pane != nil {
-		cell.Pane.Resize(w.Width, PaneContentHeight(w.Height))
-	}
-}
-
 // PaneCount returns the number of panes in the window's layout tree.
 func (w *Window) PaneCount() int {
 	count := 0
@@ -772,133 +326,6 @@ func (w *Window) mustFindPane(paneID uint32) (*LayoutCell, error) {
 		return nil, fmt.Errorf("pane %d not found in layout", paneID)
 	}
 	return cell, nil
-}
-
-func (w *Window) rootChildForPaneID(paneID uint32) (*LayoutCell, int, error) {
-	if w.IsLeadPane(paneID) {
-		return nil, -1, fmt.Errorf("cannot operate on lead pane")
-	}
-
-	root := w.logicalRoot()
-	if root == nil || root.IsLeaf() {
-		return nil, -1, fmt.Errorf("window has no root-level split")
-	}
-
-	leaf, err := w.mustFindPane(paneID)
-	if err != nil {
-		return nil, -1, err
-	}
-
-	cell := leaf
-	for cell.Parent != root {
-		cell = cell.Parent
-	}
-	return cell, cell.IndexInParent(), nil
-}
-
-// ColumnIndexForPaneID reports which top-level vertical split column contains
-// paneID. When a lead pane is anchored, it is always column 0 and the logical
-// root columns are offset to 1, 2, ...
-func (w *Window) ColumnIndexForPaneID(paneID uint32) (int, error) {
-	if w == nil || w.Root == nil {
-		return 0, fmt.Errorf("window has no layout")
-	}
-	if _, err := w.mustFindPane(paneID); err != nil {
-		return 0, err
-	}
-	if w.IsLeadPane(paneID) {
-		return 0, nil
-	}
-
-	columnBase := 0
-	root := w.logicalRoot()
-	if w.hasAnchoredLead() {
-		columnBase = 1
-	}
-	if root == nil || root.IsLeaf() || root.Dir != SplitVertical {
-		return columnBase, nil
-	}
-
-	_, idx, err := w.rootChildForPaneID(paneID)
-	if err != nil {
-		return 0, err
-	}
-	return columnBase + idx, nil
-}
-
-func (w *Window) columnContainerForPaneID(paneID uint32) (*LayoutCell, error) {
-	if w.Root == nil {
-		return nil, fmt.Errorf("window has no layout")
-	}
-	if _, err := w.mustFindPane(paneID); err != nil {
-		return nil, err
-	}
-	if w.Root.IsLeaf() || w.Root.Dir != SplitVertical {
-		return w.Root, nil
-	}
-	cell, _, err := w.rootChildForPaneID(paneID)
-	if err != nil {
-		return nil, err
-	}
-	return cell, nil
-}
-
-func firstOtherPaneID(cell *LayoutCell, exclude uint32) (uint32, bool) {
-	other := uint32(0)
-	cell.Walk(func(leaf *LayoutCell) {
-		if other != 0 || leaf == nil || leaf.Pane == nil || leaf.Pane.ID == exclude {
-			return
-		}
-		other = leaf.Pane.ID
-	})
-	return other, other != 0
-}
-
-func (w *Window) wrapColumnWithBottomPane(column *LayoutCell, pane *Pane) {
-	oldParent := column.Parent
-	oldIdx := column.IndexInParent()
-	oldX, oldY, oldW, oldH := column.X, column.Y, column.W, column.H
-	size2 := (oldH - 1) / 2
-	size1 := oldH - 1 - size2
-
-	column.ResizeSubtree(oldW, size1)
-	newLeaf := NewLeaf(pane, oldX, oldY+size1+1, oldW, size2)
-	newRoot := &LayoutCell{
-		X:        oldX,
-		Y:        oldY,
-		W:        oldW,
-		H:        oldH,
-		Dir:      SplitHorizontal,
-		Parent:   oldParent,
-		Children: []*LayoutCell{column, newLeaf},
-	}
-	column.Parent = newRoot
-	newLeaf.Parent = newRoot
-
-	if oldParent == nil {
-		w.Root = newRoot
-		return
-	}
-	oldParent.Children[oldIdx] = newRoot
-}
-
-func (w *Window) appendPaneToColumn(column *LayoutCell, pane *Pane) {
-	switch {
-	case column.IsLeaf():
-		w.wrapColumnWithBottomPane(column, pane)
-	case column.Dir == SplitHorizontal:
-		newLeaf := NewLeaf(pane, 0, 0, 0, 0)
-		newLeaf.Parent = column
-		column.Children = append(column.Children, newLeaf)
-		column.distributeEqual()
-	default:
-		w.wrapColumnWithBottomPane(column, pane)
-	}
-}
-
-func (w *Window) finishTreeMutation() {
-	w.Root.FixOffsets()
-	w.resizePTYs()
 }
 
 func reorderLayoutChildren(children []*LayoutCell, fromIdx, targetIdx int, before bool) []*LayoutCell {
@@ -1072,77 +499,6 @@ func (w *Window) movePaneWithinSplitGroup(paneID uint32, delta int) error {
 	// "before target" mode.
 	parent.Children = reorderLayoutChildren(parent.Children, idx, targetIdx, delta < 0)
 	w.finishTreeMutation()
-	return nil
-}
-
-// MovePaneToColumn reparents paneID into the logical column selected by
-// targetPaneID, appending the moved pane to the bottom of that column.
-func (w *Window) MovePaneToColumn(paneID, targetPaneID uint32) error {
-	w.assertOwner("MovePaneToColumn")
-	if col := w.leadColumn(); col != nil {
-		if (col.FindPane(paneID) != nil) != (col.FindPane(targetPaneID) != nil) {
-			return fmt.Errorf("cannot move panes across lead column")
-		}
-	}
-
-	sourceCell, err := w.mustFindPane(paneID)
-	if err != nil {
-		return err
-	}
-
-	sourcePane := sourceCell.Pane
-	sourceColumn, err := w.columnContainerForPaneID(paneID)
-	if err != nil {
-		return err
-	}
-	destColumn, err := w.columnContainerForPaneID(targetPaneID)
-	if err != nil {
-		return err
-	}
-
-	sameColumn := sourceColumn == destColumn
-	destWasRoot := destColumn == w.Root
-	anchorPaneID := targetPaneID
-	if sameColumn && paneID == targetPaneID {
-		ok := false
-		anchorPaneID, ok = firstOtherPaneID(destColumn, paneID)
-		if !ok {
-			return nil
-		}
-	}
-
-	if sameColumn || destColumn.IsLeaf() || destColumn.Dir != SplitHorizontal {
-		if destColumn.H < 2*PaneMinSize+1 {
-			return fmt.Errorf("not enough space to move pane into destination column (%d < %d)", destColumn.H, 2*PaneMinSize+1)
-		}
-	}
-
-	if w.ZoomedPaneID != 0 {
-		w.Unzoom()
-	}
-
-	sourceWasActive := w.ActivePane != nil && w.ActivePane.ID == paneID
-	if err := w.ClosePane(paneID); err != nil {
-		return err
-	}
-
-	postDestColumn := destColumn
-	if sameColumn {
-		if destWasRoot {
-			postDestColumn = w.Root
-		} else {
-			postDestColumn, err = w.columnContainerForPaneID(anchorPaneID)
-			if err != nil {
-				return err
-			}
-		}
-	}
-
-	w.appendPaneToColumn(postDestColumn, sourcePane)
-	w.finishTreeMutation()
-	if sourceWasActive {
-		w.setActive(sourcePane)
-	}
 	return nil
 }
